requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

# Project specific
torchvision
transformers
sentencepiece
pandas
omegaConf
pytorch_lightning
wandb

# Storage
dvc
dvc[gdrive]

# Testing
pytest